Add ability to decrypt PRT headers from a file
This change adds a new flag to accept a file containing prt headers to decrypt. The output of the script is a CSV file containing the results of decrypting all the prt headers in the input file.

This change will make it easier for users to validate a number of prt headers at once.

Bug: 411641206
Change-Id: I7be7976bc5b1062f7e8e3e8677e4c6f0ffc92fac